[Improve] Speed up ck file attach (#6091)
